requirements.txt
social-auth-app-django        5.0.0
social-auth-core
python3-openid
psycopg2
Pillow
djangorestframework
django-filter
Django
aiogram
